+++ demo/src/composables/sampleContent.ts
@@ -1,3 +1,8 @@
-import sampleContent from "./sampleContent.aimd?raw"
+import { ref } from 'vue'
+import sampleContentDefault from './sampleContent.aimd?raw'
 
-export const SAMPLE_AIMD = sampleContent
+export const SAMPLE_AIMD = sampleContentDefault
+
+export function useSampleContent() {
+  return ref(SAMPLE_AIMD)
+}